3단계

+++ src/main/java/com/example/scheduler/controller/ScheduleController.java
@@ -15,21 +15,23 @@ public class ScheduleController {
     @Autowired
     private ScheduleService scheduleService;
 
+    //일정 등록
     @PostMapping
-    public ScheduleDTO create(@RequestBody Map<String, String> request) {
+    public ScheduleDTO createSchedule(@RequestBody Map<String, String> request) {
         String title = request.get("title");
-        String writer = request.get("writer");
+        Long writerId = Long.parseLong(request.get("writerId"));
         String password = request.get("password");
-        return scheduleService.createSchedule(title, writer, password);
+
+        return scheduleService.createSchedule(title, writerId, password);
     }
 
     //전체 일정 조회
     @GetMapping
     public List<ScheduleDTO> getSchedules(
-            @RequestParam(required = false) String writer,
+            @RequestParam(required = false) Long writerId,
             @RequestParam(required = false) String modifiedDate
     ) {
-        return scheduleService.getSchedules(writer, modifiedDate);
+        return scheduleService.getSchedules(writerId, modifiedDate);
     }
 
     // 단건 조회
@@ -58,6 +60,11 @@ public class ScheduleController {
             @RequestBody Map<String, String> request
     ) {
         String password = request.get("password");
+
+        if (password == null || password.isBlank()) {
+            throw new IllegalArgumentException("비밀번호가 누락되었습니다.");
+        }
+
         scheduleService.deleteSchedule(id, password);
         return "삭제 성공";
     }
